Add scaffolding of 'isolate check' command
This adds the 'check' command from the original Python implementation.
Currently it just calls through to the unimplemented isolate code.

I also added a flags argument to the command line parsing so that subcommands
can specify which flags are required, and avoid having to duplicate that logic.

R=maruel@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1103653002