Stop referencing testing.T in isolatedfake.
This is needed because t.FailNow() can only be called from the main routine, as
per documentation.

Settle default namespace to "default-gzip" due to historical reason and enforce
it in fake isolated server, this will be fixed back to "default" once the old
API is deleted on the servers.

R=tandrii@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1116983002