Add counter to tracer.
tracer is still not fully functional, trace-viewer fails to display properly
when multiple eventComplete happens concurrently on a single TID when events are
not purely nested. But switching to eventBegin/eventEnd has a lot of
side-effects on the viewer, rendering both (!) Category and TIDs useless. This
will be done in a follow up which will remove both.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1152403003