batch_archive: archive item should be keyed by isolated file name
This fixes the name of the archive 'item' to be based on the name of
the isolated file rather than the isolate file. This is important since
the same isolate file may generate many different isolated outputs
(e.g. with different variable substitutions).

Additionally, we use strings.LastIndexOf to strip the extension off of the
.isolated file rather than strings.SplitN. This allows for files that have a
'.' in their name (e.g. "foo-test.1.isolated")

R=maruel@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1133123004