Add 'user friendly' datastore API.
This is a thin veneer on top of datastore.RawInterface, and may be used
with any filters, any backend, and it may be used interchangably and
inderleved with datastore.RawInterface.

In particular, this adds a convention for embedding keys inside of
loadable/savable objects.

R=vadimsh@chromium.org, dnj@chromium.org, estaab@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1259593005 .