A function to flatten nested AuthGroup.
Checks a user is in a group or not by using flat auth group
instead of nested auth groups.

This is just a beginning of series of change list to make
IsGroupMember fast.  A converted AuthDBSnapshot is stored in
memory, and in-memory cache will be updated periodically.
This flat AuthGroup will be stored to the cache.

Minor change:
- since pcg warned me buffer_test.go format error, I fixed it.

R=ukai@chromium.org

Review URL: https://codereview.chromium.org/1261263003 .