impl/memory: Make queries self-validate
This moves validation and normalization logic into the query construction
process itself. This makes it so that queries are either valid and normalized,
or they contain an error (and query building 'sticks' on the first error
encountered).

This replaces the previous build-a-questionable-query-and-then-try-to-make-it-sensible
logic (which was copied from dev_appserver for the most part).

R=dnj@google.com, dnj@chromium.org, estaab@chromium.org, tandrii@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1286943002 .