Transparently retry 404s when served from CloudEndpoints and not as json.
Endpoints occasionally serve 404 even for valid APIs, when it should have in
fact served a 503. This happens when Cloud Endpoints has an internal error. It
must be specified, in this failure mode, the AppEngine app doesn't even see the
request and there's nothing in AppEngine logs to note that this request was done
by the client, at all.

While the bug is truly inside Cloud Endpoints, work around this problem by
leveraging out automatic retry code by adding a new automatic retry trigger
based on the apparent failure mode.

This automatic trigger only happens when the response Content-Type is
text/html, not application/json. This was reported that when Endpoints has an
internal error, it will (likely) serve text/html instead of the expected
application/json. This is yet to be confirmed as this failure mode is hard to
reproduce. This is also tricky, as genuine invalid GET also return 404 with
text/html.

For now, it's the best we can do. Hopefully we'll be able to revert this change
one day.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1363883003