auth: Make LoginMode part of Authenticator state, simplify API a bit.
LoginMode has to be a part of Authenticator so that we can decide what to do
with revoked token. With OptionalLogin mode a revoked token is fine, we just
skip authentication. In other login modes this is error (user has to relogin).

This logic will be implemented in follow up CL. This one is just a refactoring
without any new functionality.

R=nodir@chromium.org
BUG=542686

Review URL: https://codereview.chromium.org/1399733004