Add simple secret store.
Similar to what components.auth provides in luci-py. Will be used to
autogenerate secret keys used in XSRF tokens and other protocols.

It is split between server/ and appengine/ because I intend to implement most
of authentication component functionality without assumption of using appengine
(so that it can be used from GCE for example). That means a persistence layer
would have to be abstracted (so that we can implement it on top of the datastore
in GAE and on top of something else in GCE). This CL is an example of how I plan
to do it.

R=iannucci@chromium.org, maruel@chromium.org, nodir@chromium.org, dnj@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1403163002