Tail-call optimization for tumble.
This is /theoretically/ not necessary, except that GAE has limits on the size
of a transaction, so this situation is practically encountered when a single
mutation needs to accomplish more than 10MB of changes within a single entity
group.

The pattern here is when a Mutation returns itself as the 'additional mutation'
to do from its RollForward method. In 'real life', the Mutation is probably
doing a query in its entity group for work to do, doing some of it, then
deferring back to tumble.

Without this optimization, tumble will end up doing 1 transaction for each
subdivided mutation, which has the potential to be very slow. With this
optimization, tumble will attempt to run any mutations for a given EG returned
by other mutations on that EG. If the transaction grows too large, the mutations
will be written in the tumble journal just like the unoptimized case.

The included test case shows this performing 5 transactions instead of 40.

DM encounters this scenario when performing the 'unblock' operation for an
Attempt where the attempt has many incoming dependencies.

This CL also fixes a bug where realMutation objects weren't generating unique
IDs on a given tumble shard processing run.

R=dnj@chromium.org, estaab@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1415383003