server/discovery: add discovery service
The discovery service accepts a list of service name and uses the
descriptions in the registry. It exposes one method Describe that
returns a list of service names and a file descriptor set that describes
the services, their message and transitive dependencies.

It is not prpc-specific

R=iannucci@chromium.org, dnj@chromium.org, vadimsh@chromium.org, estaab@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1571393006