milo: buildbot backfiller flags and de-dup
Adds the dryrun flag, which causes the backfiller to not change the datastore
if it is true. Also adds the useBuildbot flag, which allows the backfiller to
his the buildbot master directly, if CBE does not have the data.

Also de duplicates the revisions for a given repository for a build.

BUG=

Review URL: https://codereview.chromium.org/1596523006