Add support for uint8, uint16 and uint32 in luci/gae
This has come up for like 3 models in DM so far, and I'm tired of having this
be a stupid "gotcha!". I was working around it with some PropertyConverter, but
with this change it will just be obvious what the heck is going on.

They all map to positive int64 values (which should make indexes obvious), and
they include out-of-bounds checks on deserialization.

I excluded uint64, since there's no way to properly encode it in the native
datastore representation which still makes any sense for indexes.

I also added some (apparently?) missing documentation in PLS which explicitly
lays out what field types are allowed in a pls-managed struct.

R=dnj@chromium.org, estaab@chromium.org, martiniss@chromium.org, vadimsh@chromium.org
BUG=550684

Review URL: https://codereview.chromium.org/1667403003