Fix testclock race and Reset bug.
Fixes an issue where a test timer would still trigger on the first
scheduled Reset edge even if it was Reset to a later time.

Fixes a race where a timer's running state can be accessed by the
invocation goroutine.

Both of thesse were introduced during the Context channel restructuring.

Also fixes an issue with Timer where it would call the timer callback with
a negative duration if called with a negative duration. Since timer
callback is test code and usually modifies the clock, this can result in
backwards-in-time panics.

BUG=

Review URL: https://codereview.chromium.org/1758353003