LogDog: Refine service wrapper.
Refine the LogDog backend service wrapper to be less obscure and more
usable by future microservices. Update the Collector to operate with the
new service wrapper.

This also rewrites the "config changed" poll loop to be explicitly
consumed by the main thread instead of running as an unchecked
goroutine, which is not guaranteed to be scheduled according to the
Go memory model.

BUG=

Review URL: https://codereview.chromium.org/1767513002