Fix non-determinism in client/internal/tracer tests; improve readability.
The non-determinism is more apparent on Go 1.6 but the bug has always been
there. The stream is expected to not be deterministic, the events are pushed
asynchronously to not force implicitly synchronisation when using the tracer.

Put all the 'constants' and the globals into two named struct, which improves
the readability. When the file was initially 100 lines it was fine but at >400
lines, this meakes the code unreadable.

Increases back the Windows timer resolution to 1ms since it has been lowered in
Go 1.6. Ref: https://golang.org/doc/go1.6#runtime

Having a 15.6ms timer resolution makes tracing of short lived events (<10ms)
unusable. The clock resolution is only lowered when tracing is enabled, for the
duration of the trace.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1763313003