Milo: Add basic rendering tests
This adds a test in milo/frontend that feeds fake data from each endpoint to each
theme and does a render of the html to make sure that works.

This adds a new method to the themedHandler interface TestData(), which is supposed
to return a list of name/test data pairs which then gets fed into the frontend test.

go test -test.generate can generate the expectations (html files).

BUG=468043

Review URL: https://codereview.chromium.org/1774243002