LogDog: Remove storage cleanup service.
Remove the storage cleanup service. When it was initiall conceived, it
was a supplement for the fact that Google Cloud BigTable didn't support
garbage collection at the time. Recent tests suggests that it does
successfully support garbage collection, so we get the storage cleanup
for free!

R=iannucci@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1784183004