Switch the client API from hand written messages to generated messages.
The API was added in fb0578fb1a1164efb3116c9d02d2bb76858932a3.

Hacked the API to replace Content to []byte (from string) as it is supposed to.
It is a bug in Cloud Endpoints and apigen.

Changed "asdfad.nonexistent.local" to "127.0.0.1:1" as trying to resolve non
existing domain even on ".local" can take multiple seconds.

R=dnj@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1781453002