common/cloudlogging: Fix test race.
There is a race where additional enqueued entries can become buffered
waiting on test ACKs. Meanwhile the test code is waiting for a send event
before ACKing, resulting in deadlock. Change the ACK expectation to happen
after messages are consumed to avoid this case.

Note that this has no production impact. Also note that I have only seen
the deadlock on Travis and was unable to reproduce the deadlock locally, so
this is theory at the moment.

TBR=iannucci@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/1853133002