LogDog: Enable keys-only BigTable queries.
Currently, since individual log entry records are compacted together
into a RecordIO blob in BigTable, keys-only queries cannot be executed,
since the number of entries in the blob is required in order to traverse
the key space.

This modifies the BigTable schema to encode the number of records in a
row into that row's key. This enables a keys-only query to know this
value and, consequently, traverse the space without needing to load the
full data blob.

A "legacy mode" is supported with the intention of deprecating in a week
or two once uncounted BigTable data has aged off.

BUG=

Review URL: https://codereview.chromium.org/1872903002