Archivist asserts completeness through keys scan.
Currently, when the Archivist needs to archive a complete stream, it
attempts to do an archival and fails if records are missing. This is
fairly expensive, as the full archival requires Google Storage
operations and potentially large amounts of Intermediate Storage data
being pulled. A bad case is when the missing log entry is the last one
in a very long stream.

This takes advantage of the new Storage keys-only Get support to assert
completeness up-front via a keys-only scan of the log stream. This
improves archival speeds in failure cases and simplifies archival logic.

BUG=

Review URL: https://codereview.chromium.org/1874563005