impl/memory: datastore and memcache namespacing
Update GetNamespace() to return an additional boolean, which will be
true if the namespace was set and false if it was not. The AppEngine SDK
defaults to having no namespace installed ("", false), but will allow
the user to install any namespace. This allows a user to differentiate
between no namespace and an empty namespace ("", true).

Make impl/memory's datastore query operation to error if the namespace has
been set to the empty namespace (""). This reflects observed production
behavior.

BUG=

Review URL: https://codereview.chromium.org/1916463004