Switch all graph-data-creation routines to EnsureGraphData.
This removes 3 non-orthogonal APIs and replaces them with a single API which can:
  * Facilitate addition of Quests and Attempts for a user (either raw or via
    template)
  * Facilitate addition of Deps on Attempts from a currently-running Execution,
    which may, or may not create new Quests in the process (either raw or via
    template)

Additionally, the new API returns the current graph state for situations where
the 'new' graph data already exists. In particular, this feature allows
executions to be stateless: they can always ask to add 'new' dependencies and
DM will no only tell the client that the dependencies are completed, but can
also immediately transmit the results of the completed dependencies to the
execution.

This CL also adds 'template' Quest functionality to DM so that projects can
expose some pre-defined quest templates with zero or more parameters. DM will
maintain the 'BuiltBy' array for Quests so that programs can trivially verify
that a given Quest originated from a trusted template or not.

R=dnj@chromium.org, estaab@chromium.org, nodir@chromium.org
BUG=550684

Review-Url: https://codereview.chromium.org/1834873002