Actually implement Execution Activation.
This updates all tests to use randomly generated tokens, and the 'execute' test
helper to actually put the new Executions in the SCHEDULED state, instead of
skipping activation and going straight to the RUNNING state. The 'activate'
test helper is introduced to actually do the 'Activate' rpc that all Executions
would do if they were running for real (e.g. on Swarming).

This also moves all Normalization to be handled in a single `svcdec` decorator.
This way, all request normalization is handled consistently, and handler bodies
will only ever receive normalized requests.

R=dnj@chromium.org, estaab@chromium.org
BUG=550684

Review-Url: https://codereview.chromium.org/1871283002