common/lru: Add external locking, heuristic.
Add the ability for the user to control the locker instance that the
LRU uses for protection. This includes the ability to have a no-lock LRU
cache (which is obviously not goroutine-safe).

Add the ability for the user to supply the LRU heuristic callback. This
enables a user to implement external accounting callbacks or prune LRU
based on non-cache-managed factors.

BUG=

Review-Url: https://codereview.chromium.org/1960403002