Cache ErrCacheMiss errors.
Currently, the configuration cache does not cache ErrCacheMiss. This
means that if a request asks for invalid configuration paths, nothing
will be cached, and requests will flood the config service.

Update the caching byte layout to add a header "type" byte. This can
currently represent a cache hit or config.ErrCacheMiss, but can be
extended to additional errors if warranted.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/1994673004