LogDog: Add prefix registration endpoint.
Add an prefix registration API and endpoint to LogDog coordinator. This
is exposed in a separate registration module. Implementation of this
endpoint is left to a future CL.

Add Coordinator awareness of project WRITE permission. This will be checked
in the RegisterPrefix endpoint and prefix creation/access is gated on
WRITE permission.

BUG=

Review-Url: https://codereview.chromium.org/1970823005