LogDog: Implement RegisterPrefix RPC.
Implement the register prefix RPC. Remove prefix creation from the
register stream RPC. Prefix creation and access to prefix namepace is
gated on WRITE ACL for the user.

Expiration is currently not retained or enforced, and will be implemented
in a future CL.

LogDog Coordinator is no longer usable by the current Butler. A future
CL will enable the Butler to call RegisterStream appropriately.

BUG=

Review-Url: https://codereview.chromium.org/1967273002