LogDog: Implement prefix expiration.
Implement prefix space expiration. Previously, this field was present in
the protobuf, but not implemented.

This enforces that new stream creation will fail a predefined period of
time after a prefix space has been created. This value is the minimum of
the value specified in the project config, service config, and prefix
registration.

BUG=

Review-Url: https://codereview.chromium.org/1972753002