LogDog: Enable terminal index during registration.
Currently, a complete stream will result in a minimum of two Coordinator
RPCs, one to note that the stream exists and one to mark its terminal
index. For short streams where the terminal index is included in the
initial stream bundle, this can be reduced to a single RPC by allowing
the terminal index to be set during registration.

Update the Coordinator to enable the RegisterStream services RPC to
accept an optional terminal index. Update the Collector to set this and
cache it appropriately.

This additionally adds non-transient error caching to the Collector's
Coordinator cache. If a non-transient error occurs in the Collector, the
Collector will cache it.

BUG=

Review-Url: https://codereview.chromium.org/2012413003