Remove client/internal/retry.
This replaces all instances of client/internal/retry with common/retry. This
also means that isolatedclient automagically obtains the ability to be cancelled
via context.Cancel() (although this is not currently implemented).

Alters lhttp's interface a bit to simply return `func() (int, error)` and do the
retries internally. This was a bit cleaner than porting the mutable Status()
side effect idea.

R=dnj@chromium.org, maruel@chromium.org, tandrii@chromium.org, vadimsh@chromium.org
BUG=550684

Review-Url: https://codereview.chromium.org/2064283002