buildbucket client: use auth.SilentLogin
With OptionalLogin authenticator will send an anonymous request if
service account file is not present, and might send it it is expired.

The tradeoff is that if service account is not specified and there are no
cached credentials (calling by calling authutil login), buildbucket
requests will be anonymous. There are no users of buildbucket tool so
we accept this tradeoff

This might fix crbug.com/623142

R=vadimsh@chromium.org, robertocn@chromium.org
BUG=623142

Review-Url: https://codereview.chromium.org/2083083008