Update MiddlewareChain to be read-only.
Currently, MiddlewareChain is a []Middleware. The approved method of
manipulating the chain is to use slice operations (e.g., append(mc,
[]Middleware{a, b, c}). A problem arises when the base MiddlewareChain
is shared between different routers, each of which forks it in a
different direction:

base := MiddlewareChain{A, B}
r.GET(..., append(base, C))
r.GET(..., append(base, D))

The success of these operations relies on "append" reallocating "base",
which it will only do is "base" doesn't have capacity for the new
elements. If it does have capacity, the allocation will not happen, and
the first "r.GET" will end up receiving []Middleware{A, B, D}!

This CL converts MiddlewareChain to a read-only struct. All chain
operations must be performed on the struct, and will produce a new
derivative MiddlewareChain. This enforces safe manipulation of the
chain.

(Also fix a bug in "tumble" where "RequireCron" was incorrectly applied
to one of its handlers).

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2104033002