Refactor multiarg, split MGS/PLS.
Refactor the multiarg construction method to handle the various cases
inline, rather than defining and invoking methods for specific
scenarios. This enables PropertyLoadSaver-based methods to be handled
separately from MetaGetterSetter-based methods, which, in turn, enables
the two to be implemented independently.

A bit of boilerplate is removed and the odd case of MGS/PLS Map/Chan fits
better. Additionally, a few double-reflection code paths are optimized
along the way.

BUG=None

Review-Url: https://codereview.chromium.org/2048933004