Simplify the structure of LogDog bootstrapping.
This brings LogDog Butler more in line with expectations, namely that
when the bootstrapped process finishes, LogDog also finishes.

Currently, running LogDog subcommands starts in main setup code,
instantiates a Butler, runs some subcommand-specific setup code, then
runs the Butler in the common main area. This forces a rigid pattern
that employs goroutines to accommodate.

Rework this so that the actual Butler management is done in the
subcommand. This allows a much more straightforward bootstrapping
execution pattern:

1) Prepare subcommand. If anything fails here, it's cheap.
2) Create our Output (register Prefix), start Butler.
3) Run our subcommand until it terminates.
4) Flush our Butler.

This linear path is much easier to read and understand.

This also splits some of the bootstrap setup logic out of "main" and
into its own internal bootstrap library. This enables LogDog's bootstrap
environemnt to be employed by other components (e.g., Kitchen).

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2130643002