milo: swarming: make server parameter optional
This includes backward-compatible paths that will probably stay for some
time because we have buildbucket builds that have old URLs.

Also put "task/" before task id in URL because we may want to implement
bot view for swarming with links to tasks that the bot executed
recently.

Also make a few refactorings/simplifications.

R=hinoka@chromium.org, estaab@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2130003002