errors/annotate: Format reason in stack frame.
Currently, the "Reason" field is not formatted in a stack frame, causing
the text to appear like:

  reason: "Caught error: %(error)s"
  error: "The Error"

This is not very user-friendly, since the user cannot fluently read the
stack without doing variable interpolation for each line! Update this to
actually format the message. This doesn't prevent the dumping of stack
data alongside the message:

  reason: "Caught error: The Error"
  error: "The Error"

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2152933003