LogDog: Actually use default max buffer age.
The documentation for the MaxBufferAge parameter suggests that if it's
zero, a default value will be used. However, this was not actually
implemented. Historically this has not been a problem, since the only
user always supplied a non-zero value, but now that Butler is being
used elsewhere as a library, let's fix this.

TBR=hinoka@chromium.org
BUG=chromium:628321
TEST=None

Review-Url: https://codereview.chromium.org/2153383003