Add "RunMulti" to parallel.
Add RunMulti function and support structs to parallel. This is a
lightweight wrapper on top of a Runner which supports a pattern where
nested work functions share the same Runner, but are constructed and
executed in a manner that guarantees that the functions will not
deadlock.

Previously, a possible deadlock situation existed when passing a
Runner to functions that, themselves, were executed by that Runner. For
example (in pseudocode):

runner.Run(2, func(taskC) {
  taskC <- funcOne() {
    taskC <- funcOnePointOne()
    taskC <- funcOnePointTwo()
  }
  taskC <- funcTwo() {
    taskC <- funcTwoPointOne()
    taskC <- funcTwoPointTwo()
    }
})

Since there are a maximum of two processes, one possible execution order
runs "funcOne" and "funcTwo" before any inner functions are touched and,
since the Runner is now at capacity, the inner functions all block on
their parent functions' completions.

The RunMulti pattern avoids this pitfall by allowing the inner function
of a scheduled work function implicitly inherits its parent's time slot,
since the parent is blocked pending the inner function's execution. The
outer "maximum running jobs" invariant is maintained, but deadlock
situation is avoided since the parent is ensured that at least one of
its children will always be schedulable.

BUG=None
TEST=None
R=iannucci@chromium.org, nodir@chromium.org

Review-Url: https://codereview.chromium.org/2112173002