LogDog: Fix hierarchy keying.
There is a bug in LogDog hierarchy keying that encodes "0" as a numeric
component ID with 1 leading zero. It should have no leading zeroes. This
fixes that bug and adds a test for it.

There is another bug where prefix registration was causing the trailing
"+" to be registered as the stream component instead of a path
component, causing duplicate "+" components, one stream and one path, to
be registered.

Some minor cleanup is also included.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2170593002