Add "deploytool", a LUCI cloud deployment tool.
The "deploytool" is the initial iteration of the LUCI cloud service
deployment tool. It is currently capable of managing Go AppEngine and
Google Container Engine deployments.

"deploytool" loads its configuration from a series of text protobuf
files which describe configuration layout and deployment operations.
These files can be found in "common/proto/deploy".

The rest of the tool is documented in "tools/cmd/deploytool/doc.go".

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2114103002