deploytool: Allow local checkout overrides.
Sometimes it's useful to deploy against local repositories instead of
the ones defined in the layout. However, modifying layout files to
reference local repositories is burdensome, since they are versioned
files belonging to an external repository.

This adds the ability to use a local deploytool configuration file to
specify repository overrides keyed on the repository URL. The overrides
are used if the "-local" flag is passed to the "checkout" command, and
otherwise ignored.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2184453002