Add Transient and ExtractData to errors, remove MaybeLogError.
Gets rid of the awful MaybeLogError function that no-one really
liked (and nothing but DM used), in favor of annotating the error
object directly with the grpc code as soon as we can determine it,
and then using the service postlude to do all logging and conversion
to grpc's native error type.

This allows full stack information to be propagated all the way to the
top of the handler, and allows the error to remain in an 'annotated'
state for the entire stack.

R=nodir@chromium.org, vadimsh@chromium.org, dnj@chromium.org
BUG=550684

Review-Url: https://codereview.chromium.org/2223683002