LogDog: Remove use of AckID.
The Cloud Pub/Sub library no longer exports AckID. Remove references to
it. Most of these are logging.

One significant change here is the usage of the Ack ID as a unique
Pub/Sub task identifier. This was always a bit dubious, as Pub/Sub
doesn't specifically guarentee that Ack IDs won't be re-used ever.
Replace it with the current timestamp. This is also a bit dubious, since
this depends on reasonably synced clocks amongst Archivist instances.
However, since this is all part of a failure-reduction path, accepting
that as a condition is not bad.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2322673002