LogDog/Archivist: Conditionally render data.
When archiving log stream data, LogDog's Archivist supports emitting a
rendered data stream, which creates a flat file containing all of the
log stream's data. This is useful for direct consumption, but not
required. Because it doubles the amount of archival space required, it
is gated on two conditions:

- The project specifies that streams should always be rendered, or
- The stream requests to be rendered to binary by providing an
  extension.

Currently, these conditions are ignored, and the stream is always
rendered. This patch makes the Archivist actually pay attention to this
configuration and conditionally render the stream.

BUG=chromium:643243
TEST=unittest

Review-Url: https://codereview.chromium.org/2321173002