LogDog: Update web code, stream/query with auth.
Add facilities to display web viewer status, retry on missing stream,
and retry with authentication when available.

Authentication occurs as an optional asynchronous event. This wires that
event up to a retry hook that will retry the stream query and/or stream
fetching when previous attempts failed for authentication reasons. If
authentication isn't necessary, it will not be used and the first
attempt will succeed.

Additionally, update the web query code to move more logic out of the
core Polymer area and into standalone objects. The plan is to move those
objects over to TypeScript and add unit tests for their behaviors. This
should be much easier now that the objects are testable without
a full Polymer setup.

BUG=chromium:645692
TEST=local
  - Streamed logs from protected and non-protected projects.
  - Successfully logged in after failure and loaded logs.
  - Successfully queried for logs where the query required
    authentication.
  - Successfully loaded multiple log streams.

Review URL: https://codereview.chromium.org/2335223003 .