Milo: Fix pubsub ingestion not working on large builds
When a new builder gets added to a chromium waterfall, the first build contains
the commit history of the entire chromium history, which doesn't fit into
a datastore entity, even compressed.  So we just drop it, since it's not worth
the complexity to get this to work (the 2nd build will still work).

There's a bug where the err wasn't getting propogated correctly, this should fix it.

BUG=633685

Review-Url: https://codereview.chromium.org/2338373005