scheduler: More careful handling of transient errors during invocation start.
Invocations are initiated via task queue tasks. When an invocation fails to
start (i.e. move from STARTING to RUNNING or some finished state) we return
HTTP 500 to the Task Queue service, so it redelivers the same task.

This CL adds an explicit limit on number of such retries, and also makes the
state of retrying visible in the UI (as "RETRYING").

R=estaab@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2340483003