logdog_cat => "logdog", latest command, rendering.
This change renames "logdog_cat" to "logdog", the general LogDog client
CLI. It adds a "latest" command to the tool to dump the latest complete
entry in a log stream, and cleans up rendering in support of this
command.

Added a global timeout flag to "logdog" command to specify when it
should time out.

Documentation has been updated to reflect this.

In support of adding the "latest" subcommand, some changes were made to
the renderer system and its consumers:

Renderer now spceifically supports "raw" mode, where it dumps raw data
rather than attempting to make it human-readable. This is useful for
tools that want to process log data directly from the "logdog" tool.

The archive's LogStreamSource was rolled into renderer.Source, and
both are now configured to be have more like io.Reader.

BUG=None
TEST=local,unit

Review-Url: https://codereview.chromium.org/2375463002