LogDog: Split CLI tool into package and main.
Split most of the "logdog" CLI tool into a Go package called "cli".
Allow this to be called externally and accept default parameters. Wire
up a default version of the tool that supplies no default parameters.
This will function the same as the previous "logdog" tool.

The intent of this change is to have an Infra-specific version of the
tool that is built.

BUG=chromium:650684
TEST=None

Review-Url: https://codereview.chromium.org/2372703005