Add snapshotting for CIPD packages and dimensions to DM.
This makes DM automatically pin all cipd packages used by an attempt, and optionally allows the quest to specify dimensions to pin as well. This prevents confusing bugs where e.g. a certain version of kitchen is used for the first execution of an attempt, but a different version is used for a subsequent execution. This means things like the 'state' that DM passes between executions doesn't need to worry about being interpreted by different versions of the various cipd packages in the job.

Pinning dimensions could be useful for pinning instance ids (for affinity), but more trivially is useful for pinning os/cpu when pinning a VERY generic cipd spec (e.g. things with ${platform} in the spec). In this case you could potentially formulate a quest that could run on ANY platform and get consistent re-executions.

BUG=639975

Review-Url: https://codereview.chromium.org/2338153003