tsmon: respond to config file errors, test ignores
The tsmon initialization logic ignores configuration file errors. This
changes this such that the tsmon package will return an error from its
initialization function if the configuration file exists but could not
be read (e.g., corrupt, bad data, bad permissions, etc.) This aligns
with user expectations: on virtue of installing and configuring the
file, the user clearly expects the file to load. If the file fails to
load, this leads to unexpected behavior.

The file not existing will still be silently ignored, since it is an
optional file.

Also, update testing code to stub out the configuration file so that it
doesn't accidentally pick up the configuration from its host machine.
This solves a problem with the "luci-go" commit queue failing due to recent
configuration file rollouts violating testing expectations.

BUG=None
TEST=unit
R=sergeyberezin@chromium.org

Review-Url: https://codereview.chromium.org/2406783003